requirements.txt
fastapi
uvicorn
jinja2
python-multipart
sqlalchemy
openai
scipy
scikit-posthocs
seaborn
matplotlib
plotly
pandas

requests
Authlib
httpx
itsdangerous
python-dotenv
